extracted common interface methods to BaseInterface and synchronized version names with function names
