+++ fusedwake/noj/interface.py
@@ -2,77 +2,33 @@
 import fusedwake.noj.fortran as fnoj
 import fusedwake.noj.fortran_mod as fnoj_mod
 import numpy as np
+from fusedwake.interface import BaseInterface
 
 
-class NOJ(object):
+class NOJ(BaseInterface):
     # The different versions and their respective inputs
     inputs = {
         # 'py0': ['WF', 'WS', 'WD', 'K'],
         'fort_noj_av': ['x_g', 'y_g', 'z_g', 'dt', 'p_c', 'ct_c', 'ws', 'wd', 'kj',
-                  'av', 'rho', 'ws_ci', 'ws_co', 'ct_idle'],
+                        'av', 'rho', 'ws_ci', 'ws_co', 'ct_idle'],
         'fort_noj': ['x_g', 'y_g', 'z_g', 'dt', 'p_c', 'ct_c', 'ws', 'wd', 'kj',
-                  'rho', 'ws_ci', 'ws_co', 'ct_idle'],
+                     'rho', 'ws_ci', 'ws_co', 'ct_idle'],
         'fort_noj_s': ['x_g', 'y_g', 'z_g', 'dt', 'p_c', 'ct_c', 'ws', 'wd', 'kj',
-                  'rho', 'ws_ci', 'ws_co', 'ct_idle'],
+                       'rho', 'ws_ci', 'ws_co', 'ct_idle'],
         'fort_mod_noj_av': ['x_g', 'y_g', 'z_g', 'dt', 'p_c', 'ct_c', 'ws', 'wd', 'kj',
-                  'av', 'rho', 'ws_ci', 'ws_co', 'ct_idle'],
+                            'av', 'rho', 'ws_ci', 'ws_co', 'ct_idle'],
         'fort_mod_noj': ['x_g', 'y_g', 'z_g', 'dt', 'p_c', 'ct_c', 'ws', 'wd', 'kj',
-                  'rho', 'ws_ci', 'ws_co', 'ct_idle'],
+                         'rho', 'ws_ci', 'ws_co', 'ct_idle'],
         'fort_mod_noj_s': ['x_g', 'y_g', 'z_g', 'dt', 'p_c', 'ct_c', 'ws', 'wd', 'kj',
-                  'rho', 'ws_ci', 'ws_co', 'ct_idle'],
+                           'rho', 'ws_ci', 'ws_co', 'ct_idle'],
     }
     # Default variables for running the wind farm flow model
     defaults = {
         'rho': 1.225,
         'K': 0.04,
         'version': 'fort_noj_s',
-        'sup': 'quad', # ['lin' | 'quad']
+        'sup': 'quad',  # ['lin' | 'quad']
     }
-    def __init__(self, **kwargs):
-        self.set(self.defaults)
-        self.set(kwargs)
-
-    @property
-    def versions(self):
-        versions = list(self.inputs.keys())
-        versions.sort()
-        return versions
-
-    def set(self, dic):
-        """ Set the attributes of a dictionary as instance variables. Prepares
-        for the different versions of the wake model
-
-        Parameters
-        ----------
-        dic: dict
-            An input dictionary
-        """
-        for k, v in dic.items():
-            setattr(self, k, v)
-
-        # Preparing for the inputs for the fortran version
-        if 'WF' in dic:
-            self.x_g, self.y_g, self.z_g = self.WF.get_T2T_gl_coord2()
-            self.dt = self.WF.rotor_diameter
-            self.p_c = self.WF.power_curve
-            self.ct_c = self.WF.c_t_curve
-            self.ws_ci = self.WF.cut_in_wind_speed
-            self.ws_co = self.WF.cut_out_wind_speed
-            self.ct_idle = self.WF.c_t_idle
-
-    def _get_kwargs(self, version):
-        """Prepare a dictionary of inputs to be passed to the wind farm flow model
-
-        Parameters
-        ----------
-        version: str
-            The version of the wind farm flow model to run ['py0' | 'py1' | 'fort0']
-        """
-        if 'py' in version:
-            return {k:getattr(self, k) for k in self.inputs[version] if hasattr(self, k)}
-        if 'fort' in version:
-            # fortran only get lowercase inputs
-            return {(k).lower():getattr(self, k) for k in self.inputs[version] if hasattr(self, k)}
 
     def fort_noj_av(self):
         # Prepare the inputs
@@ -94,11 +50,11 @@ class NOJ(object):
             self.p_wt, self.t_wt, self.u_wt = fnoj.noj_av(**self._get_kwargs(self.version))
         except Exception as e:
             raise Exception('The fortran version {} failed with the followind inputs: {}, and the error message: {}'.format(
-                    self.version, self._get_kwargs(self.version), e))
+                self.version, self._get_kwargs(self.version), e))
         A = 0.25 * self.WF.WT.rotor_diameter**2.0
         self.c_t = self.t_wt / (0.5 * A * self.rho * self.u_wt**2.0)
         self.p_wt *= 1.0E3  # Scaling the power back to Watt
-        if len(self.ws) == 1: # We are only returning a 1D array
+        if len(self.ws) == 1:  # We are only returning a 1D array
             self.p_wt = self.p_wt[0]
             self.u_wt = self.u_wt[0]
             self.c_t = self.c_t[0]
@@ -109,21 +65,24 @@ class NOJ(object):
             self.ws = np.array([self.WS])
             self.wd = np.array([self.WD])
             self.kj = np.array([self.K])
+        else:
+            self.ws = self.WS
+            self.wd = self.WD
+            self.kj = np.full_like(self.wd, self.K)
 
         # Run the fortran code
         try:
             self.p_wt, self.t_wt, self.u_wt = fnoj.noj(**self._get_kwargs(self.version))
         except Exception as e:
             raise Exception('The fortran version {} failed with the followind inputs: {}, and the error message: {}'.format(
-                    self.version, self._get_kwargs(self.version), e))
+                self.version, self._get_kwargs(self.version), e))
         A = 0.25 * self.WF.WT.rotor_diameter**2.0
         self.c_t = self.t_wt / (0.5 * A * self.rho * self.u_wt**2.0)
         self.p_wt *= 1.0E3  # Scaling the power back to Watt
-        if len(self.ws) == 1: # We are only returning a 1D array
+        if len(self.ws) == 1:  # We are only returning a 1D array
             self.p_wt = self.p_wt[0]
             self.u_wt = self.u_wt[0]
             self.c_t = self.c_t[0]
-
 
     def fort_noj_s(self):
         self.kj = self.K
@@ -131,7 +90,7 @@ class NOJ(object):
             self.p_wt, self.t_wt, self.u_wt = fnoj.noj_s(**self._get_kwargs(self.version))
         except Exception as e:
             raise Exception('The fortran version {} failed with the followind inputs: {}, and the error message: {}'.format(
-                    self.version, self._get_kwargs(self.version), e))
+                self.version, self._get_kwargs(self.version), e))
         A = 0.25 * self.WF.WT.rotor_diameter**2.0
         self.c_t = self.t_wt / (0.5 * A * self.rho * self.u_wt**2.0)
         self.p_wt *= 1.0E3  # Scaling the power back to Watt
@@ -156,11 +115,11 @@ class NOJ(object):
             self.p_wt, self.t_wt, self.u_wt = fnoj_mod.mod_noj_av(**self._get_kwargs(self.version))
         except Exception as e:
             raise Exception('The fortran version {} failed with the followind inputs: {}, and the error message: {}'.format(
-                    self.version, self._get_kwargs(self.version), e))
+                self.version, self._get_kwargs(self.version), e))
         A = 0.25 * self.WF.WT.rotor_diameter**2.0
         self.c_t = self.t_wt / (0.5 * A * self.rho * self.u_wt**2.0)
         self.p_wt *= 1.0E3  # Scaling the power back to Watt
-        if len(self.ws) == 1: # We are only returning a 1D array
+        if len(self.ws) == 1:  # We are only returning a 1D array
             self.p_wt = self.p_wt[0]
             self.u_wt = self.u_wt[0]
             self.c_t = self.c_t[0]
@@ -177,15 +136,14 @@ class NOJ(object):
             self.p_wt, self.t_wt, self.u_wt = fnoj_mod.mod_noj(**self._get_kwargs(self.version))
         except Exception as e:
             raise Exception('The fortran version {} failed with the followind inputs: {}, and the error message: {}'.format(
-                    self.version, self._get_kwargs(self.version), e))
+                self.version, self._get_kwargs(self.version), e))
         A = 0.25 * self.WF.WT.rotor_diameter**2.0
         self.c_t = self.t_wt / (0.5 * A * self.rho * self.u_wt**2.0)
         self.p_wt *= 1.0E3  # Scaling the power back to Watt
-        if len(self.ws) == 1: # We are only returning a 1D array
+        if len(self.ws) == 1:  # We are only returning a 1D array
             self.p_wt = self.p_wt[0]
             self.u_wt = self.u_wt[0]
             self.c_t = self.c_t[0]
-
 
     def fort_mod_noj_s(self):
         self.kj = self.K
@@ -193,24 +151,13 @@ class NOJ(object):
             self.p_wt, self.t_wt, self.u_wt = fnoj_mod.mod_noj_s(**self._get_kwargs(self.version))
         except Exception as e:
             raise Exception('The fortran version {} failed with the followind inputs: {}, and the error message: {}'.format(
-                    self.version, self._get_kwargs(self.version), e))
+                self.version, self._get_kwargs(self.version), e))
         A = 0.25 * self.WF.WT.rotor_diameter**2.0
         self.c_t = self.t_wt / (0.5 * A * self.rho * self.u_wt**2.0)
         self.p_wt *= 1.0E3  # Scaling the power back to Watt
-
 
     # def python_v0(self):
     #     self.p_wt, self.u_wt, self.c_t = noj.nojarsen_v0(**self._get_kwargs(self.version))
     #
     # def python_v1(self):
     #     self.p_wt, self.u_wt, self.c_t = noj.nojarsen(**self._get_kwargs(self.version))
-
-    def __call__(self, **kwargs):
-        self.set(kwargs)
-        if hasattr(self, 'version'):
-            getattr(self, self.version)()
-            if not self.version in self.versions:
-                raise Exception("Version %s is not valid: version=[%s]"%(self.version, '|'.join(self.versions)))
-        else:
-            raise Exception("Version hasn't been set: version=[%s]"%('|'.join(self.versions)))
-        return self
